+++ src/components/ButtonGroup/index.js
@@ -1,3 +1,0 @@
-import ButtonGroup from './ButtonGroup';
-
-export { ButtonGroup };