+++ src/components/Modal/utils.js
@@ -1,0 +1,4 @@
+
+export const getTitleId = id => `${id}_title`;
+export const getContentId = id => `${id}_content`;
+