+++ stories/MediaObject.js
@@ -12,11 +12,12 @@ const sampleText = (
 );
 
 stories
-  .add('With icon left', () => (
+  .add('Default', () => (
     <MediaObject
       customTag={text('Custom HTML-tag', '') || undefined}
       figureLeft={icon}
-      flavor={array('Flavor', ['center', 'responsive'])}
+      center={boolean('Center?', false)}
+      responsive={boolean('Responsive?', true)}
       size={select('Size', ['', 'small', 'large'], 'small') || undefined}
       truncate={boolean('Truncate', false)}
       title={text('Title', 'This is a tooltip displayed on mouseover, necessary when truncate is used') || undefined}
@@ -29,7 +30,8 @@ stories
       customTag={text('Custom HTML-tag', '') || undefined}
       figureLeft={icon}
       figureRight={icon}
-      flavor={array('Flavor', ['center'])}
+      center={boolean('Center?', false)}
+      responsive={boolean('Responsive?', true)}
       size={select('Size', ['', 'small', 'large'], '') || undefined}
       truncate={boolean('Truncate', false)}
       title={text('Title', 'This is a tooltip displayed on mouseover, necessary when truncate is used') || undefined}